cambios backend

+++ prjGAS/gasbravo/views.py
@@ -31,7 +31,10 @@ def test(request):
         'all_users_count': User.objects.all().annotate(count = Count("id")),
         'total_admins': User.objects.filter(user_rol__roles = 'Admin').annotate(count = Count("id")),
         'total_clients': User.objects.filter(user_rol__roles='Cliente').annotate(count=Count("id")),
-        'all_prod': Producto.objects.all(),
+        'all_prod': Producto.objects.filter(estado = 1),
+        'total_prod': Producto.objects.filter(estado=1).annotate(count =Count("id")),
+        'total_stock': Producto.objects.filter(estado=1, cantidad_stock__gt =0).annotate(count=Count("id")),
+        'total_without_stock': Producto.objects.filter(estado=1, cantidad_stock =0).annotate(count=Count("id")),
         'user_dir': Direccion.objects.filter(user_log_id=request.session['logged_user']),
         'user': User.objects.filter(id=request.session['logged_user']),
     }
@@ -93,15 +96,21 @@ def dashboard_rol(request):
         return redirect('/')
     user_log = User.objects.get(id=request.session['logged_user'])
     context = {
-        'logged_user' : User.objects.get(id=request.session['logged_user']),
+        'logged_user': User.objects.get(id=request.session['logged_user']),
         'google_api_key': settings.GOOGLE_API_KEY,
-        'user_log': user_log,   
-        'all_bod' : Bodega.objects.all(),  
-        'all_dir' : Direccion.objects.all(),  
+        'user_log': user_log,
+        'all_bod': Bodega.objects.all(),
+        'all_dir': Direccion.objects.all(),
         'allroles': Rol.objects.all(),
-        'all_users' : User.objects.all(),   
-        'all_prod' : Producto.objects.all(),
-        'user_dir' : Direccion.objects.filter(user_log_id =request.session['logged_user'] ),
+        'all_users': User.objects.all().order_by("-created_at")[:6],
+        'all_users_count': User.objects.all().annotate(count=Count("id")),
+        'total_admins': User.objects.filter(user_rol__roles='Admin').annotate(count=Count("id")),
+        'total_clients': User.objects.filter(user_rol__roles='Cliente').annotate(count=Count("id")),
+        'all_prod': Producto.objects.filter(estado=1),
+        'total_prod': Producto.objects.filter(estado=1).annotate(count=Count("id")),
+        'total_stock': Producto.objects.filter(estado=1, cantidad_stock__gt=0).annotate(count=Count("id")),
+        'total_without_stock': Producto.objects.filter(estado=1, cantidad_stock=0).annotate(count=Count("id")),
+        'user_dir': Direccion.objects.filter(user_log_id=request.session['logged_user']),
         'user': User.objects.filter(id=request.session['logged_user']),
     }
     if user_log.user_rol.id == 2:
@@ -332,12 +341,16 @@ def create_bod(request):
             return redirect('/user/gestion_bodega')
         new_bod = Bodega.objects.create(
             prod_bod = Producto.objects.get(id = request.POST['prod_bod']),
-            guia_rem = request.POST['guia_rem'],
+            factura = request.POST['factura'],
+            proveedor = request.POST['proveedor'],
             fecha_ingr = request.POST['fecha_ingr'],
-            cantidad_stock = request.POST['cantidad_stock'],
+            cantidad= request.POST['cantidad'],
             precio_compra = request.POST['precio_compra'],
-            precio_venta = request.POST['precio_venta'],
+            
             )
+        updated_prod = Producto.objects.get(id=request.POST['prod_bod'])
+        updated_prod.cantidad_stock += int(request.POST['cantidad'])
+        updated_prod.save()
     return redirect('/user/gestion_bodega')
 
 def all_bod(request):
@@ -346,7 +359,7 @@ def all_bod(request):
         return redirect('/')
 
     context = {
-        'all_bod' : Bodega.objects.all(),
+        'all_bod': Bodega.objects.all().order_by("-fecha_ingr"),
         
     }
 
@@ -358,13 +371,20 @@ def gest_bod(request):
         'logged_user' : User.objects.get(id=request.session['logged_user']),
         'google_api_key': settings.GOOGLE_API_KEY,
         'all_bod' : Bodega.objects.all(),
-        'all_prod': Producto.objects.all(),
+        'all_prod': Producto.objects.filter(estado =1),
     }
     return render(request, 'crearbod.html', context)
 
 def delete_bod(request, number):
     borr_bod = Bodega.objects.get(id=number)
-    borr_bod.delete()
+    updated_prod = Producto.objects.filter(prod_bod = borr_bod).first()
+    if updated_prod.cantidad_stock - borr_bod.cantidad >= 0:
+        updated_prod.cantidad_stock -= borr_bod.cantidad
+        updated_prod.save()
+        borr_bod = Bodega.objects.get(id=number)
+        borr_bod.delete()
+    else:
+        messages.error(request, "No se puede eliminar el registro, generaría saldo negativo")
     return redirect('/user/gestion_bodega')
 
 #CRUD GESTION DE PRODUCTO ADMIN
@@ -383,6 +403,8 @@ def create_prod(request):
             peso = request.POST['peso'],
             tipo = request.POST['tipo'],
             color = request.POST['color'],
+            precio_venta =request.POST['precio_venta'],
+            
             )
     return redirect('/user/gestion_prod')
 
@@ -407,8 +429,13 @@ def gest_prod(request):
     return render(request, 'crearprod.html', context)
 
 def edit_prd(request, number):
+    if 'logged_user' not in request.session:
+        messages.error(request, "Please register or please log in first")
+        return redirect('/')
+    
     if request.method=='GET':
         content={
+            'logged_user': User.objects.get(id=request.session['logged_user']),
             'producto': Producto.objects.get(id=number),
             'all_prod': Producto.objects.all(),
             'all_bod' : Bodega.objects.all(),
@@ -428,10 +455,12 @@ def edit_prd(request, number):
             peso = request.POST['peso'],
             tipo = request.POST['tipo'],
             color = request.POST['color'],
+            precio_venta=request.POST['precio_venta']
 )
         return redirect('/user/gestion_prod')
 
 def delete_prod(request, number):
     borr_prod = Producto.objects.get(id=number)
-    borr_prod.delete()
+    borr_prod.estado = 0
+    borr_prod.save()
     return redirect('/user/gestion_prod')
